requirements.txt
PyFunctional
dataclasses-json
